requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
gspread==5.12.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
Pillow==10.2.0
aiofiles==23.2.1
openpyxl==3.1.2
requests==2.31.0


